% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d3Venn.R
\name{d3Venn}
\alias{d3Venn}
\title{Create a d3 Venn diagram}
\usage{
d3Venn(sets, width = NULL, height = NULL)
}
\arguments{
\item{sets}{data.frame, each row correspond to one part in the Venn diagram. Must
contain columns \code{sets} and \code{size} and may contain a column \code{label}.
Intersections are represented by a list of set names.}

\item{width}{integer, width of the htmlwidget.}

\item{height}{integer, height of the htmlwidget.}
}
\value{
an \code{htmlwidget}
}
\description{
Draw a Venn diagram given a data frame with the sizes and names of the sets.
}
\details{
To pass the size (and or label) of the intersections, the corresponding \code{sets} element
must contain a list. To create \code{list} columns you can use \code{\link[base:AsIs]{base::I()}}. See Examples.
}
\section{JavaScript Libraries}{

\itemize{
\item \url{https://github.com/upsetjs/venn.js}
\item \url{https://d3js.org/}
}
}

\examples{
# Venn diagram with 2 disjoint sets both of equal size
d3Venn(data.frame(sets = c("A", "B"), size = c(20, 20)))

# Venn diagram with 2 disjoint sets with differnt sizes and labels
d3Venn(data.frame(sets = c("A", "B"), size = c(10, 20), label = c("Set 1", "Set 2")))

# Venn diagram with 2 overlapping sets
d3Venn(data.frame(sets = I(list("A", "B", list("A", "B"))),
                  size = c(10, 10, 2)))

# Venn diagram where B is a subset of A
d3Venn(data.frame(sets = I(list("A", "B", list("A", "B"))),
                  size = c(20, 10, 10)))
}
\seealso{
\code{\link[htmlwidgets:createWidget]{htmlwidgets::createWidget()}}
}
